advbox_family/ODD/requirements-gpu.txt
opencv-python==4.4.0.44
future==0.18.2
xmltodict==0.12.0
tqdm==4.50.1
tensorflow-addons==0.11.2
easydict
tensorflow-gpu==2.3.0
numpy
nibabel